fix: change lastReleaseDateDayCount from Int to Double
The XYZRank API returns lastReleaseDateDayCount as a floating point number
(e.g., 1.9794993171296296), not an integer. This was causing JSON parsing
errors. Fixed by changing the field type from Int to Double in both
XYZRankEpisode and XYZRankPodcast models.

+++ composeApp/src/commonMain/kotlin/com/opoojkk/podium/data/model/xyzrank/XYZRankModels.kt
@@ -23,7 +23,7 @@ data class XYZRankEpisode(
     val primaryGenreName: String,
     val totalEpisodesCount: Int,
     val openRate: Double,
-    val lastReleaseDateDayCount: Int
+    val lastReleaseDateDayCount: Double
 )
 
 @Serializable
@@ -53,7 +53,7 @@ data class XYZRankPodcast(
     val authorsText: String,
     val trackCount: Int,
     val lastReleaseDate: String,
-    val lastReleaseDateDayCount: Int,
+    val lastReleaseDateDayCount: Double,
     val firstEpisodePostTime: String,
     val activeRate: Double,
     val avgDuration: Double,
